Update target to es5
When using a higher target parcel tries to grab a "regeneratorRuntime"
from babel. As far as I can tell, there is no actual fix for this that
doesn't require installing babel of some kind. But I don't feel like
doing that, so I'm doing this instead.

In the future, it might be worth just removing parcel and going back to
webpack...